requirement.txt
matplotlib==3.7.2
numpy==1.25.2
opencv-python==4.5.5.62
Pillow==10.0.0
progress==1.6
scipy==1.11.1
tensorboardX==2.6.2.2
torch==2.0.1
torchvision==0.15.2